refactor: reviews minor change
